requirements.txt
aiosmtplib==3.0.2
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
build==1.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
greenlet==3.2.4
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
limits==5.5.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pip-tools==7.5.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
